+++ google-cloud/fedora-coreos/kubernetes/variables.tf
@@ -46,9 +46,17 @@ variable "worker_type" {
   default     = "n1-standard-1"
 }
 
+variable "os_stream" {
+  type        = string
+  description = "Fedora CoreOS stream for compute instances (e.g. stable, testing, next)"
+  default = "stable"
+}
+
+# Deprecated
 variable "os_image" {
   type        = string
   description = "Fedora CoreOS image for compute instances (e.g. fedora-coreos)"
+  default = ""
 }
 
 variable "disk_size" {