+++ google-cloud/fedora-coreos/kubernetes/workers.tf
@@ -8,6 +8,7 @@ module "workers" {
   network      = google_compute_network.network.name
   worker_count = var.worker_count
   machine_type = var.worker_type
+  os_stream    = var.os_stream
   os_image     = var.os_image
   disk_size    = var.disk_size
   preemptible  = var.worker_preemptible